+++ test/grpc/transport/http2_test.exs
@@ -4,6 +4,7 @@ defmodule GRPC.Transport.HTTP2Test do
   alias GRPC.Transport.HTTP2
 
   @channel %Channel{scheme: "http", host: "grpc.io"}
+  %Stream{channel: @channel, path: "/foo/bar"}
   alias GRPC.Client.Stream
 
   defp assert_header({key, _v} = pair, headers) do
